ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mediapipe==0.10.35
opencv-python-headless
numpy
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
